experiments/pyproject.toml
[project]
name = "experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "accelerate>=0.20.0",
    "bitsandbytes>=0.41.0",
    "pillow>=9.0.0",
    "torch>=2.1.0",
    "tqdm>=4.65.0",
    "transformers==4.51.3",
]
